lambda-deployment/requirements.txt
boto3==1.40.11
python-dotenv==1.1.1
requests==2.32.4